requirements.txt
duckdb
python-dotenv
openai
fastapi
uvicorn